+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Newsletter page template with all meta field settings
+- Auto list all newsletters as signup banner options
+
+### Changed
+
+- Newsletter page auto lists all child pages with signup forms
+
+### Deprecated
+
+- The Pick and The Cortado single page templates
+- Hardcoded newsletter signup banner options for The Pick and The Cortado
+
 ## [3.8.1] - 2022-11-21
 
 ### Added