+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Support Section default and url code alternate values can now set one off as the default
+- CMB2 meta field basic validation. Can validate required and max words via data attributes
 
 ### Removed